+++ packages/manager/apps/pci-object-storage/src/pages/objects/container/object/delete/Delete.page.tsx
@@ -1,11 +1,13 @@
 import { ApiError } from '@ovh-ux/manager-core-api';
 import { DeletionModal } from '@ovh-ux/manager-pci-common';
 import { useNotifications } from '@ovh-ux/manager-react-components';
-import { OdsText } from '@ovhcloud/ods-components/react';
+import { OdsInput, OdsMessage, OdsText } from '@ovhcloud/ods-components/react';
 import { Translation, useTranslation } from 'react-i18next';
 import { useNavigate, useParams, useSearchParams } from 'react-router-dom';
+import { useEffect, useMemo, useState } from 'react';
 import { useDeleteObject } from '@/api/hooks/useObject';
 import { useStorage } from '@/api/hooks/useStorages';
+import { PERMANENTLY_DELETE_MSG } from '@/constants';
 
 export default function DeletePage() {
   const { addSuccess, addError } = useNotifications();
@@ -16,6 +18,31 @@ export default function DeletePage() {
   const decodedObjectName = objectName.replace(/~2F/g, '/');
   const [searchParams] = useSearchParams();
   const region = searchParams.get('region');
+  const versionId = searchParams.get('versionId');
+
+  const isLocalZone = searchParams.get('isLocalZone');
+  const isSwift = searchParams.get('isSwift');
+  const isVersioningDisabled = searchParams.get('isVersioningDisabled');
+  const isVersioningSuspended = searchParams.get('isVersioningSuspended');
+
+  const [deleteConfirmation, setDeleteConfirmation] = useState(null);
+  const [isTouched, setIsTouched] = useState(false);
+  const [isValid, setIsValid] = useState(false);
+
+  useEffect(() => {
+    if (deleteConfirmation === PERMANENTLY_DELETE_MSG) {
+      setIsValid(true);
+    } else {
+      setIsValid(false);
+    }
+  }, [deleteConfirmation]);
+
+  let nameError: string | undefined;
+  useEffect(() => {
+    if (isTouched && !isValid) {
+      nameError = t('pci-common:common_field_error_pattern');
+    }
+  }, [isTouched, isValid]);
 
   const { storage, isPending: isPendingStorage } = useStorage(
     projectId,
@@ -23,7 +50,23 @@ export default function DeletePage() {
     region,
   );
 
-  const goBack = () => navigate(`../?region=${region}`);
+  const isPermanentDelete = useMemo(() => {
+    return (
+      versionId ||
+      isLocalZone ||
+      isSwift ||
+      isVersioningDisabled ||
+      isVersioningSuspended
+    );
+  }, [
+    versionId,
+    isLocalZone,
+    isSwift,
+    isVersioningDisabled,
+    isVersioningSuspended,
+  ]);
+
+  const goBack = () => navigate(`../?region=${region}&refetch=true`);
 
   const onCancel = goBack;
   const onClose = goBack;
@@ -33,6 +76,7 @@ export default function DeletePage() {
     objectName: decodedObjectName,
     storage,
     region,
+    versionId,
     onError(error: ApiError) {
       addError(
         <Translation ns="objects/delete">
@@ -78,13 +122,23 @@ export default function DeletePage() {
   return (
     <DeletionModal
       title={t(
-        'pci_projects_project_storages_containers_container_object_delete_object_title',
+        isPermanentDelete
+          ? 'pci_projects_project_storages_containers_container_object_delete_object_permanently_title'
+          : 'pci_projects_project_storages_containers_container_object_delete_object_title',
       )}
       onCancel={onCancel}
       onClose={onClose}
       onConfirm={onConfirm}
       isPending={isPending}
-      isDisabled={isPending}
+      isDisabled={
+        isPending ||
+        ((versionId ||
+          isLocalZone ||
+          isSwift ||
+          isVersioningDisabled ||
+          isVersioningSuspended) &&
+          !isValid)
+      }
       submitText={t(
         'pci_projects_project_storages_containers_container_object_delete_submit_label',
       )}
@@ -92,12 +146,61 @@ export default function DeletePage() {
         'pci_projects_project_storages_containers_container_object_delete_cancel_label',
       )}
     >
-      <OdsText>
-        {t(
-          'pci_projects_project_storages_containers_container_object_delete_content',
-          { object: decodedObjectName },
-        )}
-      </OdsText>
+      {isPermanentDelete ? (
+        <>
+          <OdsMessage
+            color="warning"
+            className="mt-6 mb-6"
+            isDismissible={false}
+          >
+            <OdsText>
+              {t(
+                isLocalZone ||
+                  isSwift ||
+                  isVersioningDisabled ||
+                  isVersioningSuspended
+                  ? 'pci_projects_project_storages_containers_container_object_delete_permanent_delete_no_version_text'
+                  : 'pci_projects_project_storages_containers_container_object_delete_permanent_delete_version_text',
+                { object: decodedObjectName },
+              )}
+            </OdsText>
+          </OdsMessage>
+
+          <div className="block w-full">
+            <label
+              slot="label"
+              className={`block w-full max-w-2xl text-[#4d5592] font-semibold ${
+                nameError ? 'text-critical' : ''
+              }`}
+            >
+              {t(
+                'pci_projects_project_storages_containers_container_object_delete_permanent_delete_help',
+                { confirmKey: PERMANENTLY_DELETE_MSG },
+              )}
+            </label>
+            <OdsInput
+              className="block w-full mt-3"
+              value={deleteConfirmation}
+              name="deleteConfirmation"
+              color="primary"
+              isRequired
+              onOdsChange={(event) => {
+                setDeleteConfirmation(`${event.detail.value || ''}`);
+              }}
+              onOdsBlur={() => {
+                setIsTouched(true);
+              }}
+            />
+          </div>
+        </>
+      ) : (
+        <OdsText>
+          {t(
+            'pci_projects_project_storages_containers_container_object_delete_content',
+            { object: decodedObjectName },
+          )}
+        </OdsText>
+      )}
     </DeletionModal>
   );
 }