+++ packages/manager/modules/nutanix/src/dashboard/component/add-nodes-modal/component.js
@@ -1,0 +1,18 @@
+import template from './template.html';
+import controller from './controller';
+
+export default {
+  bindings: {
+    goBack: '<',
+    nodeTechnicalDetails: '<',
+    cluster: '<',
+    server: '<',
+    nodePricing: '<',
+    nodeOrderLinkGenerator: '<',
+    expressOrderLink: '<',
+    handleSuccess: '<',
+    isMultipleNodesOrderEnabled: '<',
+  },
+  template,
+  controller,
+};