+++ packages/manager/modules/nutanix/src/dashboard/component/node-action-menu/component.js
@@ -1,0 +1,19 @@
+import template from './template.html';
+import controller from './controller';
+
+export default {
+  bindings: {
+    serviceName: '<',
+    server: '<',
+    hasOnlyMinimumNode: '<',
+    initialCommitmentSize: '<',
+    onPowerOn: '&',
+    onPowerOff: '&',
+    onInstall: '&',
+    onReinstall: '&',
+    onUninstall: '&',
+    onTerminate: '&',
+  },
+  template,
+  controller,
+};